chore: Merge more rbac files
- Remove cache.go -> authz.go
- Remove query.go -> authz.go
- Remove role.go -> roles.go
